polish: explicit fp32 mask cast; spy-limitation and shape-math comments
Final whole-branch review items: the attention mask now casts to fp32
explicitly (island must not depend on upstream dtype), and two test
comments document the stream spy's PatchExpand blind spot and the
local-blocks test's shape math.

+++ src/heal_swin_nnx/models/healswin.py
@@ -213,8 +213,9 @@ class WindowAttention(nnx.Module):
 
         if mask is not None:
             nW = mask.shape[0]
+            # explicit fp32: the logits island must not depend on attn's dtype upstream
             attn = (attn.reshape(B_ // nW, nW, self.num_heads, N, N)
-                    + mask.astype(attn.dtype)[None, :, None])
+                    + mask.astype(jnp.float32)[None, :, None])
             attn = attn.reshape(-1, self.num_heads, N, N)
         attn = jax.nn.softmax(attn, axis=-1).astype(self.dtype)  # island exit
         attn = self.attn_drop(attn)
